requirements.txt
moviepy==2.0.0dev2
selenium==4.6.0
python-dotenv==0.21.0
selenium-stealth==1.0.6
pocketbase==0.8.0
typeguard==3.0.2
streamlit==1.23.1 
pylint
async-generator==1.10
attrs==21.4.0
certifi==2022.5.18.1
cffi
charset-normalizer==2.0.12
cryptography==37.0.2
h11==0.13
idna==3.3
outcome==1.1.0
pycparser==2.21
pycryptodome==3.14.1
pyOpenSSL==22.0.0
PySocks==1.7.1
requests==2.27.1
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
wsproto==1.1.0
aiohttp
